Revert "Waitingminigamesquasher"
