+++ EscapeApp/app/src/main/java/com/context/kroket/escapeapp/mainscreens/GameOver.java
@@ -13,9 +13,6 @@ public class GameOver extends AppCompatActivity {
         setContentView(R.layout.activity_game_over);
     }
 
-    /**
-     * Disables going back to the previous activity.
-     */
     @Override
     public void onBackPressed() {
     }
