+++ EscapeApp/app/src/main/java/com/context/kroket/escapeapp/mainscreens/GameWon.java
@@ -13,9 +13,6 @@ public class GameWon extends AppCompatActivity {
         setContentView(R.layout.activity_game_won);
     }
 
-    /**
-     * Disables going back to the previous activity.
-     */
     @Override
     public void onBackPressed() {
     }
