+++ EscapeApp/app/src/main/java/com/context/kroket/escapeapp/minigames/A_CodeCrackerCodeview.java
@@ -135,13 +135,6 @@ public class A_CodeCrackerCodeview extends AppCompatActivity {
     }
   }
 
-  /**
-   * Disables going back to the previous activity.
-   */
-  @Override
-  public void onBackPressed() {
-  }
-
   // ******************************************************//
   // *********** ONLY USED FOR TESTING PURPOSES ***********//
   // ******************************************************//
