+++ EscapeApp/app/src/main/java/com/context/kroket/escapeapp/minigames/A_Code_Cracker_Pictureview.java
@@ -49,11 +49,4 @@ public class A_Code_Cracker_Pictureview extends AppCompatActivity {
     ((ActivityManager) this.getApplicationContext()).setCurrentActivity(this);
   }
 
-  /**
-   * Disables going back to the previous activity.
-   */
-  @Override
-  public void onBackPressed() {
-  }
-
 }
