+++ EscapeApp/app/src/main/java/com/context/kroket/escapeapp/minigames/C_ColorSequence.java
@@ -176,11 +176,4 @@ public class C_ColorSequence extends AppCompatActivity {
   public static int getRunthroughs() {
     return runtroughs;
   }
-
-  /**
-   * Disables going back to the previous activity.
-   */
-  @Override
-  public void onBackPressed() {
-  }
 }
