+++ EscapeApp/app/src/main/java/com/context/kroket/escapeapp/minigames/D_Lock.java
@@ -203,11 +203,4 @@ public class D_Lock extends AppCompatActivity {
         // Change the current activity to this
         ((ActivityManager) this.getApplicationContext()).setCurrentActivity(this);
     }
-
-    /**
-     * Disables going back to the previous activity.
-     */
-    @Override
-    public void onBackPressed() {
-    }
 }
