+++ EscapeApp/app/src/main/res/layout/e__squasher.xml
@@ -8,29 +8,24 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:background="@drawable/woodwall"
-    tools:context=".minigames.Waiting_Squasher">
+    tools:context="com.context.kroket.escapeapp.minigames.E_Squasher">
 
-    <CheckedTextView android:id="@+id/squashtext"
+    <TextView android:id="@+id/squashtext"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:hint="@string/squashtext"
         android:layout_gravity="center_horizontal"
-        android:textColor="#f7f7f7"
         android:layout_weight="0.07"
         android:textStyle="bold"
         android:textSize="20dp"
-        android:gravity="center"
-        android:checked="false" />
+        android:gravity="center" />
 
     <ImageButton
         android:id="@+id/bugbutton"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:background="@android:color/transparent"
-        android:src="@drawable/bug"
-        android:layout_centerVertical="true"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentStart="true" />
+        android:src="@drawable/bug" />
